% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_functions.R
\name{CalcScores}
\alias{CalcScores}
\title{Projection onto a basis}
\usage{
CalcScores(data, basis, weightinv = NULL)
}
\arguments{
\item{data}{Data matrix to be projected, where each column is a representation on the original field}

\item{basis}{Basis matrix}

\item{weightinv}{If NULL, uses standard SVD projection. Otherwise, uses weighted projection.}
}
\value{
Matrix of basis coefficients
}
\description{
Calculates the coefficients given by projecting data onto a basis
}
\examples{
# First generate some data

l <- 100 # dimension of output
n <- 10 # number of runs
DataBasis <- MakeDataBasis(data = matrix(runif(l*n), nrow=l, ncol=n), RemoveMean = TRUE) # data is 100x10

# Project the (centred) ensemble onto the first 3 vectors of the SVD basis

Coefficients <- CalcScores(data = DataBasis$CentredField, basis = DataBasis$tBasis[,1:3])

# Instead of projecting using W = I, define a W with varying diagonal

W <- runif(l, 1, 5) * diag(l) # 100x100 diagonal matrix
W_inv <- GetInverse(W) # inverse needed for projection
Coefficients_weighted <- CalcScores(data = DataBasis$CentredField, basis = DataBasis$tBasis[,1:3], weightinv = W_inv)

}
