% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_functions.R
\name{ExplainT}
\alias{ExplainT}
\title{Number of basis vectors required to explain proportion of data}
\usage{
ExplainT(DataBasis, vtot = 0.95, weightinv = NULL)
}
\arguments{
\item{vtot}{The total proportion of variability in the data to be explained by the truncated basis}

\item{weightinv}{The inverse of W}

\item{basis}{Basis matrix}

\item{data}{Data matrix}
}
\value{
The number of basis vectors required to explain vtot of the data.
}
\description{
Finds the truncated basis that explains a set proportion of the variaiblity in the data.
}
