% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_functions.R
\name{fn}
\alias{fn}
\title{Idealised function with spatial output}
\usage{
fn(x)
}
\arguments{
\item{x}{A vector of 6 values, corresponding to x_1, ..., x_6.}
}
\value{
A 10x10 field.
}
\description{
A 6 parameter toy function that gives output over a 10x10 field.
}
\examples{
n <- 60
sample <- as.data.frame(2*maximinLHS(n,6) - 1)
colnames(sample) <- c("x1","x2","x3","x4","x5","x6")
output <- array(c(rep(0,100*n)), dim=c(10,10,n))
for (i in 1:n){
  output[,,i] <- fn(as.numeric(sample[i,]))
}
dim(output) <- c(100, n) # vectorising spatial output, so each column of the data matrix is a single realisation of the function

}
