% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_functions.R
\name{GetInverse}
\alias{GetInverse}
\title{Matrix inversion via cholesky decomposition}
\usage{
GetInverse(W)
}
\arguments{
\item{W}{square positive definite variance matrix}
}
\value{
Inverse of W, with attributes 'identity' and 'diagonal', used by other functions in the package to make calculations more efficient.
}
\description{
Inverts matrix W, assigning attributes for whether W is diagonal, to speed up other calculations.
}
\examples{
Winv <- GetInverse(diag(100))
attributes(Winv) # diagonal = TRUE, identity = TRUE

Winv2 <- GetInverse(runif(100,0.1,1)*diag(100))
attributes(Winv2) # diagonal = TRUE, identity = FALSE

Winv3 <- GetInverse(seq(0.1,1,length=100) \%*\% t(seq(0.1,1,length=100)) + 0.1*diag(100))
attributes(Winv3) # diagonal = FALSE, identity = FALSE

}
