% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_functions.R
\name{MakeDataBasis}
\alias{MakeDataBasis}
\title{Formatting data}
\usage{
MakeDataBasis(data, weightinv = NULL, RemoveMean = TRUE)
}
\arguments{
\item{data}{a matrix containing individual fields in the columns (i.e. the matrix has dimension lxn)}

\item{weightinv}{the inverse of lxl positive definite weight matrix W. If NULL, the identity matrix is used}

\item{RemoveMean}{if TRUE, centres the data prior to calculating the basis}
}
\value{
\item{tBasis}{The (weighted) SVD basis of the centred ensemble if RemoveMean = TRUE, of the original data otherwise}
\item{CentredField}{The centred data if RemoveMean = TRUE, the original data otherwise.}
\item{EnsembleMean}{The mean across the columns of the data. A zero vector if RemoveMean = FALSE}
}
\description{
Formats data so that it is in the correct form for use in other functions, and calculates the (weighted) SVD basis of the ensemble
}
