% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_functions.R
\name{Recon}
\alias{Recon}
\title{Field reconstructions from coefficients}
\usage{
Recon(coeffs, basis)
}
\arguments{
\item{coeffs}{Coefficient vector}

\item{basis}{Basis matrix}
}
\value{
Reconstructed field.
}
\description{
Given a vector of coefficients for a basis, calculates the field
}
