% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_functions.R
\name{ReconBasis}
\alias{ReconBasis}
\title{Calculate a set of vectors from weights and a basis}
\usage{
ReconBasis(weights, basis)
}
\arguments{
\item{weights}{A vector of weights}

\item{basis}{Original basis}
}
\value{
New basis vector(s)
}
\description{
Give basis vectors from linear combinations
}
