% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_functions.R
\name{ReconError}
\alias{ReconError}
\title{Reconstruction error}
\usage{
ReconError(obs, basis, weightinv = NULL, scale = TRUE)
}
\arguments{
\item{obs}{The observations}

\item{basis}{Basis to project and reconstruct the observations with}

\item{weightinv}{Inverse of weight matrix W. If NULL (default), calculates the mean squared error}

\item{scale}{If TRUE, scales by the dimension (so analogous to mean squared error)}
}
\value{
The reconstruction error
}
\description{
Calculates the reconstruction error, R_W(basis, obs), of the observations given a basis and W.
}
