% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_functions.R
\name{ReconObs}
\alias{ReconObs}
\title{Project and reconstruct a given field}
\usage{
ReconObs(obs, basis, ...)
}
\arguments{
\item{obs}{Vector over original field}

\item{basis}{Basis matrix}
}
\value{
Reconstruction of the original field.
}
\description{
Gives the reconstruction of a field using a basis, by projecting and back-projecting on this basis.
}
\examples{

}
