% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_functions.R
\name{ResidBasis}
\alias{ResidBasis}
\title{Find the residual basis}
\usage{
ResidBasis(basisvectors, data, weightinv = NULL, ...)
}
\arguments{
\item{basisvectors}{Basis vector(s) to project the data onto}

\item{data}{Data matrix for to find a basis for}
}
\value{
The full basis for the data, i.e. the basis vectors passed into the function, with the residual basis vectors appended
}
\description{
Given basis vectors and data, calculate the residual basis to complete the basis for the data
}
