% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_functions.R
\name{VarExplained}
\alias{VarExplained}
\title{Calculating the proportion of data explained by a basis}
\usage{
VarExplained(basis, data, weightinv = NULL)
}
\arguments{
\item{basis}{The basis}

\item{data}{The data to be explained}
}
\value{
The proportion of variability in the data that is explained by the basis
}
\description{
Calculates the proportion of the data that is explained by projection onto a basis.
}
