% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_functions.R
\name{VarMSEplot}
\alias{VarMSEplot}
\title{Produce a VarMSEplot}
\usage{
VarMSEplot(DataBasis, obs, RecVarData = NULL, weightinv = NULL,
  min.line = TRUE, bound = TRUE, qmax = NULL, ...)
}
\description{
Calculates and plots the reconstruction error and proportion of variability explained for each truncated basis
}
\details{
@param DataBasis A DataBasis object, containing the data and basis to be used to produce the VarMSEplot
 @param obs Observation vector
 @param RecVarData if given, the output given by a previous call of VarMSEplot, so that plots can be reproduced without rerunning potentially slow calculations
 @param weightinv Inverse of W used for calculating the reconstruction error. If NULL, calculates the mean squared error
 @param min.line If TRUE, plots a solid horizontal line at the minimum value of the reconstruction error
 @param bound If TRUE, plots a dotted horizontal line at the value of the history matching bound
 @param qmax if not NULL, value for the maximum size of the truncated basis to consider. Useful if W is non-diagonal, and either n or l is large

@return A VarMSEplot, and a matrix containing the plotted data, arranged in columns as (Reconstruction error, proportion of variability explained)

@export
}
