% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_functions.R
\name{VarProj}
\alias{VarProj}
\title{Matrix projection}
\usage{
VarProj(mat, basis, weightinv = NULL)
}
\arguments{
\item{mat}{A square matrix to be projected onto the basis}

\item{basis}{The basis to project with}

\item{weightinv}{The inverse of positive definite matrix W. If NULL, uses the standard projection, otherwise projects in the norm given by W.}
}
\value{
The projection of the original matrix on the basis.
}
\description{
Projects a variance matrix onto a given basis
}
