% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_functions.R
\name{WeightOptim}
\alias{WeightOptim}
\title{Function used within optimiser for minimisation of the reconstruction error for rotated basis, subject to constraints}
\usage{
WeightOptim(x, basis, obs, data, weightinv, v = 0.1, newvectors = NULL)
}
\arguments{
\item{x}{Vector giving the linear combination of the basis to use}

\item{basis}{The basis that is being rotated}

\item{obs}{Observation vector}

\item{data}{Ensemble data}

\item{weightinv}{Inverse of matrix W}

\item{v}{The proportion of variability to be explained by the basis vector}

\item{newvectors}{If the reconstruction error should account for any previous basis vectors}
}
\value{
The reconstruction error
}
\description{
Given a vector of weights, gives the new basis vector as a linear combination of the original basis, and calculates the reconstruction error, subject to a variability constraint
}
