% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_functions.R
\name{wsvd}
\alias{wsvd}
\title{Weighted singular value decomposition}
\usage{
wsvd(data, weightinv = NULL)
}
\arguments{
\item{data}{n x l matrix to calculate basis from (i.e. rows are output fields).}

\item{weightinv}{l x l inverse of W. If NULL, calculates standard SVD.}
}
\value{
The weighted SVD of the data.
}
\description{
Calculates the SVD basis across the output, given the inverse of W.
}
